// types/app.d.ts
export type PathLike = string;
// ...
export type GitStatus =
  | 'modified'
  | 'ignored'
  | 'unmodified'
  | '*modified'
  | '*deleted'
  | '*added'
  | 'absent'
  | 'deleted'
  | 'added'
  | '*unmodified'
  | '*absent'
  | '*undeleted'
  | '*undeletemodified'
  | 'unmerged';
// ...
export type BranchStatus = 'clean' | 'unstaged' | 'uncommitted' | 'unmerged';